added tel contact field
In accordance with the security.txt RFC it is also allowed to have one or more phone numbers listed. As per chapter `2.5.3 Contact`

 2.5.3. Contact

The "Contact" field indicates a method that researchers should use for reporting security vulnerabilities such as an email address, a phone number, and/or a web page with contact information. This field MUST always be present in a "security.txt" file. If this field indicates a web URI, then it MUST begin with "https://" (as per Section 2.7.2 of [RFC7230]). Security email addresses should use the conventions defined in Section 4 of [RFC2142].
The value MUST follow the URI syntax described in Section 3 of [RFC3986]. This means that "mailto" and "tel" URI schemes must be used when specifying email addresses and telephone numbers, as defined in [RFC6068] and [RFC3966].