Feature/redirect check (#76)
* redirect check after request

After the request is made we will check the response history to see if any redirects have happened. If this is the case we will check the current url and check if the path remains the same. If this is not the case than we will stop processing the request.